orginized services, fixed file upload and added avatar thumbnail
