+++ src/modules/relation/Relation.js
@@ -2,7 +2,9 @@ import { Add, ArrowBack, Send } from "@mui/icons-material";
 import { Avatar, Box, Button, Dialog, Grid, Paper, Stack, TextField, Toolbar, Typography } from "@mui/material";
 import { useCallback, useEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
-import { createOrUpdateRelation, getAllLessons, methods } from "../../services/urls";
+import { getAllLessons } from "../../services/lessons_services";
+import { createOrUpdateRelation } from "../../services/relations_services";
+import { methods } from "../../services/urls";
 import { homePath } from "../../utils/paths";
 import { stringAvatar } from "../../utils/strings";
 import { getUserId, getUserUuid } from "../../utils/user";
